Spotify tracks dataset

A data.frame that is for analyzing and visualizing. It is a dataset of Spotify songs with different genres and their audio features. The raw data is downloaded from Kaggle and cleaned before use.

Usage

Spotify_dataset

Format

A data.frame with 20 columns, which are:

Track_id

The Spotify ID for the track

Artists

The artists' names who performed the track. If there is more than one artist, they are separated by a ;

Album_name

The album name in which the track appears

Track_name

Name of the track

Explicit

Whether or not the track has explicit lyrics (true = yes it does; false = no it does not OR unknown)

Key

The key the track is in. Integers map to pitches using standard Pitch Class notation. E.g. 0 = C, 1 = C♯/D♭, 2 = D, and so on. If no key was detected, the value is -1

Mode

Mode indicates the modality (major or minor) of a track, the type of scale from which its melodic content is derived. Major is represented by 1 and minor is 0

Time_signature

An estimated time signature. The time signature (meter) is a notational convention to specify how many beats are in each bar (or measure). The time signature ranges from 3 to 7 indicating time signatures of 3/4, to 7/4.

Track_genre

The genre in which the track belongs

Popularity

The popularity of a track is a value between 0 and 10, with 10 being the most popular

Duration_ms

This value describes the track length. It is between 0 and 10, which is scaled from the track length in milliseconds

Danceability

Danceability describes how suitable a track is for dancing. A value of 0 is least danceable and 10 is most danceable

Energy

A measure from 0 to 10 and represents a perceptual measure of intensity and activity. Typically, energetic tracks feel fast, loud, and noisy.

Loudness

The loudness of a track is a value between 0 and 10, with 10 in largest decibels (dB)

Speechiness

Speechiness detects the presence of spoken words in a track. The value id from 0 to 1. The more exclusively speech-like the recording (e.g. talk show, audio book, poetry), the closer to 10 the attribute value

Acousticness

A confidence measure from 0 to 10 of whether the track is acoustic. 10 represents high confidence the track is acoustic

Instrumentalness

Predicts whether a track contains no vocals. The closer the instrumentalness value is to 10, the greater likelihood the track contains no vocal content

Liveness

Detects the presence of an audience in the recording. The value is from 0 to 10. Higher liveness values represent an increased probability that the track was performed live.

Valence

A measure from 0 to 10 describing the musical positiveness conveyed by a track. Tracks with high valence sound more positive (e.g. happy, cheerful, euphoric), while tracks with low valence sound more negative (e.g. sad, depressed, angry)

Tempo

A measurement of the speed or pace of a given piece. The value is from 0 to 10. The higher value the speed is faster.

Source

<https://www.kaggle.com/datasets/maharshipandya/-spotify-tracks-dataset?resource=download>