Calculate the score of Spotify tracks

Calculate the score of Spotify tracks according to the features and the level for each feature that user choose.

Usage

Spotify_scoring(genres, features, levels)

Arguments

genres

A string contain the name of the genres that the user chooses.

features

A list containing 5 characters. Each character represents the feature that the user choose

levels

A list containing 5 numerics. Each numeric represent the value of each feature that user want.

Value

list_values list containing the following attributes:

result

containing the top10 scores tracks

user_genre

contain the genre that user choose

user_feature

contain the features that user choose

user_level

contain the level of each feature that user choose

Author

Ting Yang

Examples

Spotify_scoring(genres = "Funk",
features=c("Danceability","Energy","Loudness","Speechiness","Liveness"),
levels = c(9,8,5,4,7))
#> $result
#>                                                               Artists
#> 437                                                  DJ Ghost Floripa
#> 707                                                    Wilson Pickett
#> 269                                                       James Brown
#> 805                                              Snoop Dogg;The Doors
#> 265                                                     George Benson
#> 464 Snoop Dogg;Surjeet Singh;Akshay Kumar;Kuldeep Ral;Manjeet Ral;RDB
#> 294                                                              Kyan
#> 103                                                     2Pac;Big Syke
#> 688                                      MC Marks;Mc Davi;Mc Don Juan
#> 625                                         MC Teteu;JC NO BEAT;DJ F7
#>                                Track_name Track_genre Danceability Energy
#> 437              Mega Funk Reload Vuk Vuk        Funk     9.106599   8.13
#> 707                           Let Me Know        Funk     8.395939   7.29
#> 269                Get Up Offa That Thing        Funk     8.395939   7.76
#> 805 Riders On The Storm - Fredwreck Remix        Funk     8.913706   7.86
#> 265    Give Me the Night (Single Version)        Funk     8.548223   8.01
#> 464           Singh Is Kinng - Title Song        Funk     9.035533   8.13
#> 294                              Mandrake        Funk     8.974619   7.70
#> 103         All Eyez On Me (ft. Big Syke)        Funk     8.852792   7.23
#> 688                          Ela Tá Tá Tá        Funk     8.903553   7.21
#> 625                 Ta Com Saudade de Mim        Funk     8.883249   7.75
#>     Loudness Speechiness Liveness  score_1 score_2  score_3  score_4 score_5
#> 437 8.175647   1.5440415     2.36 9.893401    9.87 6.824353 7.544041    5.36
#> 707 7.289089   1.3782383     6.35 9.395939    9.29 7.710911 7.378238    9.35
#> 269 7.776483   3.6787565     5.45 9.395939    9.76 7.223517 9.678756    8.45
#> 805 8.186375   1.5751295     1.22 9.913706    9.86 6.813625 7.575130    4.22
#> 265 7.976250   0.6331606     3.87 9.548223    9.99 7.023750 6.633161    6.87
#> 464 8.337125   0.7761658     1.24 9.964467    9.87 6.662875 6.776166    4.24
#> 294 8.328617   2.3937824     1.07 9.974619    9.70 6.671383 8.393782    4.07
#> 103 8.057821   2.5595855     3.27 9.852792    9.23 6.942179 8.559585    6.27
#> 688 7.970886   0.4818653     2.85 9.903553    9.21 7.029114 6.481865    5.85
#> 625 8.106653   0.5709845     1.07 9.883249    9.75 6.893347 6.570984    4.07
#>     final_score
#> 437    8.287794
#> 707    8.253889
#> 269    8.252785
#> 805    8.237648
#> 265    8.235506
#> 464    8.235426
#> 294    8.215455
#> 103    8.212819
#> 688    8.205462
#> 625    8.194836
#> 
#> $user_genre
#> [1] "Funk"
#> 
#> $user_feature
#> [1] "Danceability" "Energy"       "Loudness"     "Speechiness"  "Liveness"    
#> 
#> $user_level
#> [1] 9 8 5 4 7
#>