radar plot for songs

Draw a radar plot to show the value of one song (rank) of each feature and also show the level of each feature that user choose.

Usage

radar_plot(x, rank, ...)

Arguments

x

is an object of class spotify.

rank

is the rank of the top 10 songs that we want to plot

...

other possible parameters for this method.

Value

fig is the radar plot

Author

Ting Yang

Examples

top10_song <- Spotify_scoring(genres = "Funk",
features=c("Danceability","Energy","Loudness","Speechiness","Liveness"),
levels = c(9,8,5,4,7))
radar_plot(top10_song, 1)