Calculate the similarity of Spotify tracks

Calculate the score of Spotify tracks according to the features and the level for each feature that user choose.

Usage

spotify_similarity(genres, features, levels)

Arguments

genres

A string contain the name of the genres that the user chooses.

features

A list containing 5 characters. Each character represents the feature that the user choose

levels

A list containing 5 numerics. Each numeric represent the value of each feature that user want.

Value

list_values list containing the following attributes:

result

containing the top10 similarity tracks

user_genre

contain the genre that user choose

user_feature

contain the features that user choose

user_level

contain the level of each feature that user choose

Author

Ting Yang

Examples

spotify_similarity(genres = "Funk",
features=c("Danceability","Energy","Loudness","Speechiness","Liveness"),
levels = c(9,8,5,4,7))
#> $result
#>                              Artists                              Track_name
#> 4                             Anitta                                Envolver
#> 1                        Coolio;L.V.                      Gangsta's Paradise
#> 9            Anitta;Chencho Corleone                                    Gata
#> 5               Anitta;Missy Elliott                                   Lobby
#> 7               Anitta;Missy Elliott                                   Lobby
#> 3               Anitta;Missy Elliott                                   Lobby
#> 2  Clean Bandit;Sean Paul;Anne-Marie Rockabye (feat. Sean Paul & Anne-Marie)
#> 8                    Anitta;Swae Lee                                 Poquito
#> 6                             Anitta                          Boys Don't Cry
#> 10                   Anitta;Swae Lee                                 Poquito
#>    Track_genre Danceability Energy Loudness Speechiness Liveness
#> 4         Funk     8.253807   7.33 8.159740   0.8777202    0.909
#> 1         Funk     6.568528   5.14 7.302776   0.6145078    3.980
#> 9         Funk     5.258883   7.99 8.363391   1.2746114    7.870
#> 5         Funk     7.705584   7.11 8.200063   0.3979275    1.430
#> 7         Funk     7.705584   7.11 8.200063   0.3979275    1.430
#> 3         Funk     7.705584   7.11 8.200063   0.3979275    1.430
#> 2         Funk     7.309645   7.63 8.409263   0.5419689    1.800
#> 8         Funk     7.116751   8.51 8.416292   1.2953368    3.090
#> 6         Funk     6.030457   8.74 8.496199   0.4010363    1.480
#> 10        Funk     7.116751   8.51 8.416292   1.2953368    3.090
#> 
#> $user_genre
#> [1] "Funk"
#> 
#> $user_feature
#> [1] "Danceability" "Energy"       "Loudness"     "Speechiness"  "Liveness"    
#> 
#> $user_level
#> [1] 9 8 5 4 7
#>